requirements.txt
Django==2.2.4
django-phonenumber-field==3.0.1
djangorestframework==3.10.2
phonenumbers==8.10.17
pytz==2019.2
Babel==2.7.0
Shapely==1.6.4.post2
